Unit Test fixes and some reorg
Moved the timer & interval based tests to the correct fixture and fixed
abug in the extension method which wasn't providing the scheduler to the
Timer when creating it
